SOLVED THE PUZZLE!
So the thing is; previous commit before solves all the test cases, and
this version gives me 100% on the solution; now one must traverse the
labyrinth a lot (like 500 turns) in order to construct a valid path.
This solution is a bit "random" in the way that Kirk selects the random
direction (I choose right for submission). However I consider this
solved in order to move on to funnier projects (raytracing!!!) and it is
a bit tiring to write code in codingame ide. (Also job is demanding and
I want to relax a bit more